Se modifico un bug de Histograma

+++ proc-de-imagenes/src/main/java/untref/service/HistogramService.java
@@ -2,7 +2,7 @@ package untref.service;
 
 public interface HistogramService {
 
-	int[] getArrayHistogram();
+	double[] getArrayHistogram();
 
 	void BarChartDraw();
